requirements.txt
pygame
python-chess
numpy
torch
matplotlib
